+++ compiler/lib-wasm/binaryen.mli
@@ -16,11 +16,17 @@
  * Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA 02111-1307, USA.
  *)
 
+type link_input =
+  { module_name : string
+  ; file : string
+  }
+
 val link :
-     runtime_files:string list
-  -> input_files:string list
+     ?options:string list
+  -> inputs:link_input list
   -> opt_output_sourcemap:string option
   -> output_file:string
+  -> unit
   -> unit
 
 val dead_code_elimination :
@@ -33,8 +39,10 @@ val dead_code_elimination :
 
 val optimize :
      profile:Driver.profile option
+  -> ?options:string list
   -> opt_input_sourcemap:string option
   -> input_file:string
   -> opt_output_sourcemap:string option
   -> output_file:string
+  -> unit
   -> unit